Fix clear behavior for Search page
