+++ src/components/SearchBar/SearchIcon.tsx
@@ -1,0 +1,19 @@
+export default function SearchIcon({ fill }: { fill: string }) {
+  return (
+    <svg
+      width="23"
+      height="23"
+      viewBox="0 0 23 23"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <g id="Search">
+        <path
+          id="Path 786"
+          d="M22.0795 19.8562L17.7524 15.5303C19.1996 13.5402 19.8433 11.0776 19.5552 8.63384C19.267 6.19007 18.0682 3.94492 16.1977 2.34599C14.3273 0.747073 11.9228 -0.0881246 9.46397 0.00737169C7.00512 0.102868 4.67265 1.12196 2.93186 2.86113C1.19108 4.6003 0.16995 6.93167 0.0721814 9.39043C-0.0255874 11.8492 0.807204 14.2546 2.4044 16.1265C4.00159 17.9984 6.24581 19.1994 8.68931 19.4898C11.1328 19.7802 13.596 19.1388 15.5875 17.6934L19.8584 21.9644C19.9969 22.1102 20.1626 22.2274 20.3463 22.3091C20.53 22.3909 20.728 22.4357 20.929 22.441C21.13 22.4462 21.33 22.4117 21.5178 22.3396C21.7055 22.2676 21.8771 22.1592 22.0229 22.0208C22.1688 21.8824 22.2859 21.7167 22.3677 21.533C22.4494 21.3493 22.4943 21.1512 22.4995 20.9502C22.5047 20.7493 22.4703 20.5493 22.3982 20.3616C22.3262 20.1739 22.2179 20.0021 22.0795 19.8562ZM16.5564 9.78545C16.5564 11.1133 16.1627 12.4113 15.425 13.5154C14.6873 14.6195 13.6388 15.4799 12.412 15.9881C11.1853 16.4963 9.8353 16.6294 8.53294 16.3704C7.23058 16.1114 6.03427 15.4721 5.09526 14.5333C4.15625 13.5944 3.51672 12.3981 3.25755 11.0958C2.99837 9.79343 3.13119 8.44344 3.6392 7.2166C4.14721 5.98976 5.00763 4.94128 6.11161 4.20342C7.21559 3.46555 8.51355 3.07153 9.84141 3.07134C11.6214 3.07345 13.3279 3.78143 14.5866 5.04009C15.8453 6.29875 16.5534 8.00543 16.5555 9.78545H16.5564Z"
+          fill={fill}
+        />
+      </g>
+    </svg>
+  )
+}
